Initial commit
Things working:
Initialisation of Lua state, closing of Lua state, checking the type of a Lua
variable (more as a tryout), execution of arbitrary strings in Lua (note, not
"safe" per se).
Work laid out for JValue userdata type in Lua for storing J binary blobs (as
returned by 3!:1), but untested so far.